Improve metadata management

+++ peer.go
@@ -90,7 +90,6 @@ func NewPeerState(conn net.Conn) *peerState {
 }
 
 func (p *peerState) Close() {
-	//log.Println("Closing connection to", p.address)
 	p.conn.Close()
 	// No need to close p.writeChan. Further writes to p.conn will just fail.
 }
@@ -151,7 +150,7 @@ func (p *peerState) SendBitfield(bs *Bitset) {
 }
 
 func (p *peerState) SendExtensions(supportedExtensions map[int]string,
-	metadataSize int) {
+	metadataSize int64) {
 
 	handshake := ExtensionHandshake{
 		M:            make(map[string]int, len(supportedExtensions)),
